requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
anyio==4.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cattrs==23.1.2
certifi==2022.12.7
charset-normalizer==3.2.0
ciso8601==2.3.0
click==8.1.6
contourpy==1.0.6
cycler==0.11.0
exceptiongroup==1.1.2
Flask==2.3.2
fonttools==4.38.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
kiwisolver==1.4.4
MarkupSafe==2.1.3
packaging==22.0
paho-mqtt==1.6.1
Pillow==9.3.0
platformdirs==3.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2022.7
reactivex==4.0.4
requests==2.31.0
requests-cache==1.1.0
soupsieve==2.4.1
telegram==0.0.1
thefuzz==0.19.0
threadpoolctl==3.1.0
timple==0.1.6
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.16

